+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.18.0-beta.1 (Unreleased)
 - Added support for service version 2023-05-03 and 2023-08-03.
+- Added RehydratePendingToCold value to ArchiveStatus enum.
 
 ## 12.17.0 (2023-07-11)
 - Includes all features from 12.17.0-beta.1.